Do not export functions
This may change with future versions of Morloc, but for now exports are
restricted to top-level scripts and are used to determine the commands
that are exported to the user.

I do eventually need a way to select what functions in a module are
imported into another module. As it stands, the contents of a module are
a bit unclear. Every function within the scope of all sourced
files/languages are technically available (though not easy to find). All
functions with declared type signatures are available.

I need something like the Haskell module system. Or something better, I
need to look over existing paradigms. One thing I need to add is
namespaces. Also, I need to decide how to handle name conflicts. I could
do masking (like in R), though I think raising an error would be better.