Use ints in fibonacci
I could use size_t. I can write packers to translate that to ints in the
json. And I could write a UInt type for R and other languages. But
ultimately, all I am doing is doubling the size of the number that can
be expressed. Hardly a solution. Since the fibonacci function is mostly
just for use in simple examples, I'll stick to simple ints for now.